resources/[05]-Barcode-Generator/[05]-Barcode-Generator/requirements.txt
dnspython==2.6.1
lxml==5.2.2
Pillow==9.5.0
pymongo==4.8.0
python-barcode==0.13.1
python-docx==0.8.11
